1、支持真实版面样例迭代 2、版面编码暂时采用“4位原始版面序号+生成时间”的方式命名 3、结束录音的操作改为单击，避免书写有被逼迫感 4、录音提示音改为语音提示

+++ app/src/main/java/com/example/xmatenotes/DotInfoActivity.java
@@ -43,7 +43,7 @@ public class DotInfoActivity extends BaseActivity {
 //	private int xMin, xMax, yMin, yMax;//刻画当前笔划的坐标范围
 	private Rect rect = null;
 	private int xD,yD;//当前笔迹点坐标
-	private int page = XmateNotesApplication.DEFAULT_INT;//当前页码
+	private long page = XmateNotesApplication.DEFAULT_INT;//当前页码
 	private float time = XmateNotesApplication.DEFAULT_FLOAT;//存储当前视频碎片时刻信息
 	private int videoID = XmateNotesApplication.DEFAULT_INT;//存储当前视频碎片ID信息
 	private byte penID = XmateNotesApplication.DEFAULT_INT;//存储当前笔ID
