+++ app/src/main/java/com/example/xmatenotes/ui/BaseActivity.java
@@ -104,12 +104,12 @@ public class BaseActivity extends AppCompatActivity {
      * @param x 命令第一个点的横坐标
      * @param y 命令第一个点的纵坐标
      */
-    public void receiveRecognizeResult(Gesture ges, int pageID, int x, int y){
+    public void receiveRecognizeResult(Gesture ges, long pageID, int x, int y){
         Log.e(TAG,"BaseActivity: receiveRecognizeResult()");
     }
 
     //普通书写延迟响应回调方法
-    public void writeTimerFinished(int pageID, int x, int y){
+    public void writeTimerFinished(long pageID, int x, int y){
 
     }
